componentize LMG life goals
